requirements.txt
tqdm
numpy
Pillow
scikit-learn
matplotlib